Update README.md
The information in the README.md file was a bit confusing.
So I updated the file with a better explanation of how the
game actually works. I also mentioned a link in the README
file that takes the reader to the game website to have a
look and feel of the game!